+++ aiosmtpd/tests/test_handlers.py
@@ -4,14 +4,13 @@ import asyncio
 import unittest
 
 from aiosmtpd.controller import Controller
-from aiosmtpd.handlers import (
-    AsyncMessage, Debugging, Mailbox, Message, Proxy, Sink)
+from aiosmtpd.handlers import AsyncMessage, Debugging, Mailbox, Proxy, Sink
 from aiosmtpd.smtp import SMTP as Server
 from contextlib import ExitStack
 from io import StringIO
 from mailbox import Maildir
 from operator import itemgetter
-from smtplib import SMTP, SMTPRecipientsRefused
+from smtplib import SMTP, SMTPDataError, SMTPRecipientsRefused
 from tempfile import TemporaryDirectory
 from unittest.mock import call, patch
 
@@ -152,22 +151,25 @@ Testing
 """.format(peer))
 
 
+class DataHandler:
+    def __init__(self):
+        self.content = None
+        self.original_content = None
+
+    @asyncio.coroutine
+    def handle_DATA(self, server, session, envelope):
+        self.content = envelope.content
+        self.original_content = envelope.original_content
+        return '250 OK'
+
+
 class TestMessage(unittest.TestCase):
-    def setUp(self):
-        self.handled_message = None
-
-        class MessageHandler(Message):
-            def handle_message(handler_self, message):
-                self.handled_message = message
-
-        self.handler = MessageHandler()
-
     def test_message(self):
-        # In this test, the message data comes in as bytes.
-        controller = Controller(self.handler)
+        # In this test, the message content comes in as a bytes.
+        handler = DataHandler()
+        controller = Controller(handler)
         controller.start()
         self.addCleanup(controller.stop)
-
         with SMTP(controller.hostname, controller.port) as client:
             client.sendmail('anne@example.com', ['bart@example.com'], """\
 From: Anne Person <anne@example.com>
@@ -177,22 +179,17 @@ Message-ID: <ant>
 
 Testing
 """)
-        self.assertEqual(self.handled_message['subject'], 'A test')
-        self.assertEqual(self.handled_message['message-id'], '<ant>')
-        self.assertIsNotNone(self.handled_message['X-Peer'])
-        self.assertEqual(
-            self.handled_message['X-MailFrom'], 'anne@example.com')
-        self.assertEqual(self.handled_message['X-RcptTo'], 'bart@example.com')
+        # The content is not converted, so it's bytes.
+        self.assertEqual(handler.content, handler.original_content)
+        self.assertIsInstance(handler.content, bytes)
+        self.assertIsInstance(handler.original_content, bytes)
 
     def test_message_decoded(self):
-        # With a server that decodes the data, the messages come in as
-        # strings.  There's no difference in the message seen by the
-        # handler's handle_message() method, but internally this gives full
-        # coverage.
-        controller = DecodingController(self.handler)
+        # In this test, the message content comes in as a string.
+        handler = DataHandler()
+        controller = DecodingController(handler)
         controller.start()
         self.addCleanup(controller.stop)
-
         with SMTP(controller.hostname, controller.port) as client:
             client.sendmail('anne@example.com', ['bart@example.com'], """\
 From: Anne Person <anne@example.com>
@@ -202,12 +199,9 @@ Message-ID: <ant>
 
 Testing
 """)
-        self.assertEqual(self.handled_message['subject'], 'A test')
-        self.assertEqual(self.handled_message['message-id'], '<ant>')
-        self.assertIsNotNone(self.handled_message['X-Peer'])
-        self.assertEqual(
-            self.handled_message['X-MailFrom'], 'anne@example.com')
-        self.assertEqual(self.handled_message['X-RcptTo'], 'bart@example.com')
+        self.assertNotEqual(handler.content, handler.original_content)
+        self.assertIsInstance(handler.content, str)
+        self.assertIsInstance(handler.original_content, bytes)
 
 
 class TestAsyncMessage(unittest.TestCase):
@@ -481,6 +475,124 @@ Testing
                 )
 
 
+class HELOHandler:
+    @asyncio.coroutine
+    def handle_HELO(self, server, session, envelope, hostname):
+        return '250 geddy.example.com'
+
+
+class EHLOHandler:
+    @asyncio.coroutine
+    def handle_EHLO(self, server, session, envelope, hostname):
+        return '250 alex.example.com'
+
+
+class MAILHandler:
+    @asyncio.coroutine
+    def handle_MAIL(self, server, session, envelope, address, options):
+        envelope.mail_options.extend(options)
+        return '250 Yeah, sure'
+
+
+class RCPTHandler:
+    @asyncio.coroutine
+    def handle_RCPT(self, server, session, envelope, address, options):
+        envelope.rcpt_options.extend(options)
+        if address == 'bart@example.com':
+            return '550 Rejected'
+        envelope.rcpt_tos.append(address)
+        return '250 OK'
+
+
+class DATAHandler:
+    @asyncio.coroutine
+    def handle_DATA(self, server, session, envelope):
+        return '599 Not today'
+
+
+class NoHooksHandler:
+    pass
+
+
+class TestHooks(unittest.TestCase):
+    def test_rcpt_hook(self):
+        controller = Controller(RCPTHandler())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            with self.assertRaises(SMTPRecipientsRefused) as cm:
+                client.sendmail('anne@example.com', ['bart@example.com'], """\
+From: anne@example.com
+To: bart@example.com
+Subject: Test
+
+""")
+            self.assertEqual(cm.exception.recipients, {
+                'bart@example.com': (550, b'Rejected'),
+                })
+
+    def test_helo_hook(self):
+        controller = Controller(HELOHandler())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            code, response = client.helo('me')
+        self.assertEqual(code, 250)
+        self.assertEqual(response, b'geddy.example.com')
+
+    def test_ehlo_hook(self):
+        controller = Controller(EHLOHandler())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            code, response = client.ehlo('me')
+        self.assertEqual(code, 250)
+        lines = response.decode('utf-8').splitlines()
+        self.assertEqual(lines[-1], 'alex.example.com')
+
+    def test_mail_hook(self):
+        controller = Controller(MAILHandler())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            client.helo('me')
+            code, response = client.mail('anne@example.com')
+        self.assertEqual(code, 250)
+        self.assertEqual(response, b'Yeah, sure')
+
+    def test_data_hook(self):
+        controller = Controller(DATAHandler())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            with self.assertRaises(SMTPDataError) as cm:
+                client.sendmail('anne@example.com', ['bart@example.com'], """\
+From: anne@example.com
+To: bart@example.com
+Subject: Test
+
+Yikes
+""")
+            self.assertEqual(cm.exception.smtp_code, 599)
+            self.assertEqual(cm.exception.smtp_error, b'Not today')
+
+    def test_no_hooks(self):
+        controller = Controller(NoHooksHandler())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            client.helo('me')
+            client.mail('anne@example.com')
+            client.rcpt(['bart@example.com'])
+            code, response = client.data("""\
+From: anne@example.com
+To: bart@example.com
+Subject: Test
+
+""")
+            self.assertEqual(code, 250)
+
+
 class CapturingServer(Server):
     def __init__(self, *args, **kws):
         self.warnings = None
@@ -510,6 +622,38 @@ class AsyncDeprecatedHandler:
         pass
 
 
+class DeprecatedHookServer(Server):
+    def __init__(self, *args, **kws):
+        self.warnings = None
+        super().__init__(*args, **kws)
+
+    @asyncio.coroutine
+    def smtp_EHLO(self, arg):
+        with patch('aiosmtpd.smtp.warn') as mock:
+            yield from super().smtp_EHLO(arg)
+        self.warnings = mock.call_args_list
+
+    @asyncio.coroutine
+    def smtp_RSET(self, arg):
+        with patch('aiosmtpd.smtp.warn') as mock:
+            yield from super().smtp_RSET(arg)
+        self.warnings = mock.call_args_list
+
+    @asyncio.coroutine
+    def ehlo_hook(self):
+        pass
+
+    @asyncio.coroutine
+    def rset_hook(self):
+        pass
+
+
+class DeprecatedHookController(Controller):
+    def factory(self):
+        self.smtpd = DeprecatedHookServer(self.handler)
+        return self.smtpd
+
+
 class TestDeprecation(unittest.TestCase):
     # handler.process_message() is deprecated.
     def test_deprecation(self):
@@ -527,7 +671,7 @@ Testing
         self.assertEqual(len(controller.smtpd.warnings), 1)
         self.assertEqual(
             controller.smtpd.warnings[0],
-            call('Use handler.handle_DATA instead of .process_message()',
+            call('Use handler.handle_DATA() instead of .process_message()',
                  DeprecationWarning))
 
     def test_deprecation_async(self):
@@ -545,5 +689,29 @@ Testing
         self.assertEqual(len(controller.smtpd.warnings), 1)
         self.assertEqual(
             controller.smtpd.warnings[0],
-            call('Use handler.handle_DATA instead of .process_message()',
+            call('Use handler.handle_DATA() instead of .process_message()',
+                 DeprecationWarning))
+
+    def test_ehlo_hook_deprecation(self):
+        controller = DeprecatedHookController(Sink())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            client.ehlo('example.com')
+        self.assertEqual(len(controller.smtpd.warnings), 1)
+        self.assertEqual(
+            controller.smtpd.warnings[0],
+            call('Use handler.handle_EHLO() instead of .ehlo_hook()',
+                 DeprecationWarning))
+
+    def test_rset_hook_deprecation(self):
+        controller = DeprecatedHookController(Sink())
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            client.rset()
+        self.assertEqual(len(controller.smtpd.warnings), 1)
+        self.assertEqual(
+            controller.smtpd.warnings[0],
+            call('Use handler.handle_RSET() instead of .rset_hook()',
                  DeprecationWarning))